Features and Maintenance + Relation SQL
Apartments have been updated to link to a Features and Maintenance page.
Features and Maintenance link back to Apartments, and displays their respective data alongside the apartments it's associated with.

Added SQL code that can be run to associate Apartments with features and maintenance. Makes two tables: Apar_Feat and Apar_Main.